refactor: extract a2a data models and TaskStore into separate modules
Move 16 Pydantic model classes to synapse/a2a_models.py and TaskStore
to synapse/task_store.py. Re-exports in a2a_compat.py maintain backward
compatibility for all existing imports.
